rename: auth.schedule 패키지를 auth.scheduler로 변경
- 타 도메인 패키지 명칭과의 일관성을 위해 패키지 이름 수정
- RefreshTokenCleanupTask를 새로운 패키지로 이동
- 패키지 변경에 따른 관련 테스트 코드 내 import 문 업데이트
- 기능적 변경 없음; 프로젝트 구조의 가독성 및 유지보수성 개선
